- #20: Restructure before adding indexedDB
