+++ src/hanzhong/components/HanzhongContext.tsx
@@ -2,8 +2,8 @@ import { createContext } from 'react';
 
 import type { HanzhongBonusType, HanzhongType } from '../types';
 import { HANZHONG_DATA } from '../data';
+import type { HanzhongUserDataType } from '../persistence/hanzhong-user-data-type';
 import { useSafeContext } from '../../components/utils/use-safe-context';
-import type { HanzhongUserDataType } from '../../persistence/hanzhong-user-data-type';
 
 export type HanzhongContextType = {
   hanzhong: HanzhongType;
