+++ src/hanzhong/components/utils/get-user-data-value-by-key.ts
@@ -1,4 +1,4 @@
-import type { HanzhongUserDataType } from '../../../persistence/hanzhong-user-data-type';
+import type { HanzhongUserDataType } from '../../persistence/hanzhong-user-data-type';
 
 export const getUserDataValueByKey = (userData: HanzhongUserDataType, key: string): number => {
   return userData[key] ?? 0;
