+++ src/hanzhong/utils/calculate-special-trainings-bonuses.ts
@@ -1,7 +1,7 @@
 import { calculateSpecialTrainingBonus } from './calculate-special-training-bonus';
 import { HANZHONG_TECH_IDS } from '../constants/items-ids';
 import type { HanzhongBonusType, HanzhongType } from '../types';
-import type { HanzhongUserDataType } from '../../persistence/hanzhong-user-data-type';
+import type { HanzhongUserDataType } from '../persistence/hanzhong-user-data-type';
 
 export type SpecialTrainingsBonusesType = {
   vanguardCamp: HanzhongBonusType;
