--- src/persistence/user-data-type.ts
@@ -1,5 +1,0 @@
-import type { HanzhongUserDataType } from './hanzhong-user-data-type';
-
-export type UserDataType = {
-  hanzhong: HanzhongUserDataType;
-};
