--- src/persistence/users-data-type.ts
@@ -1,3 +1,0 @@
-import type { UserDataType } from './user-data-type';
-
-export type UsersDataType = Record<string, UserDataType>;
